Add files via upload
In this part  we will be performing Data Collection and Pre-Processing from the ChEMBL Database. 
We are going to collect original data form the ChEMBL Database.  
ChEMBL Database is a database that contains curated bioactivity data more than 2million compaunds so it going to be the same data that researchers on the field are collecting